pyproject.toml
[project]
name = "evaluate-model"
version = "0.1.0"
description = "A Python-based framework for evaluating Large Language Models (LLMs) across different test suites"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.14.0",
    "tqdm>=4.67.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "anthropic>=0.75.0",
    "click>=8.3.1",
]

[dependency-groups]
dev = [
    "pylint>=4.0.4",
    "astroid>=4.0.2",
    "isort>=7.0.0",
    "ruff>=0.14.11",
]

